+++ OmopTransformer/SUS/AE/Observation/SourceOfReferralForAE/SusAESourceOfReferralForAE.cs
@@ -2,14 +2,14 @@
 using OmopTransformer.Omop.Observation;
 using OmopTransformer.Transformation;
 
-namespace OmopTransformer.SUS.APC.Observation.SourceOfReferralForOutpatients;
+namespace OmopTransformer.SUS.AE.Observation.SourceOfReferralForAE;
 
 [Notes(
     "Notes",
     "Observations do not require a standardized test or other activity to generate clinical fact. Typical observations are medical history, family history, lifestyle choices, healthcare utilization patterns, social circumstances etc",
     "Valid Observation Concepts are not enforced to be from any domain.  They should still be standard concepts and typically belong to the Observation or Measurement domain.",
     "Observations can be stored as attribute value pairs, with the attribute as the Observation Concept and the value representing the clinical fact. This fact can be stored as a Concept (value_as_concept), a numerical value (value_as_number) or a verbatim string (value_as_string)")]
-internal class SusAPCSourceOfReferralForOutpatients : OmopObservation<SusAPCSourceOfReferralForOutpatientsRecord>
+internal class SusAESourceOfReferralForAE : OmopObservation<SusAESourceOfReferralForAERecord>
 {
     [CopyValue(nameof(Source.NHSNumber))]
     public override string? nhs_number { get; set; }
@@ -20,15 +20,15 @@ internal class SusAPCSourceOfReferralForOutpatients : OmopObservation<SusAPCSour
     [ConstantValue(4258129, "Referral by")]
     public override int? observation_concept_id { get; set; }
 
-    [Transform(typeof(DateConverter), nameof(Source.StartDateHospitalProviderSpell))]
+    [Transform(typeof(DateConverter), nameof(Source.ArrivalDate))]
     public override DateTime? observation_date { get; set; }
 
-    [Transform(typeof(DateAndTimeCombiner), nameof(Source.StartDateHospitalProviderSpell), nameof(Source.StartTimeHospitalProviderSpell))]
+    [Transform(typeof(DateAndTimeCombiner), nameof(Source.ArrivalDate), nameof(Source.ArrivalTime))]
     public override DateTime? observation_datetime { get; set; }
 
     [ConstantValue(32818, "EHR administration record")]
     public override int? observation_type_concept_id { get; set; }
 
-    [CopyValue(nameof(Source.ReferrerCode))]
+    [CopyValue(nameof(Source.SourceofReferralForAE))]
     public override string? value_as_string { get; set; }
 }
