+++ OmopTransformer/SUS/AE/Observation/SourceOfReferralForAE/SusAESourceOfReferralForAERecord.cs
@@ -1,11 +1,11 @@
 ﻿using OmopTransformer.Annotations;
 
-namespace OmopTransformer.SUS.AE.Observation.SourceOfReferralForOutpatients;
+namespace OmopTransformer.SUS.AE.Observation.SourceOfReferralForAE;
 
 [DataOrigin("SUS")]
-[Description("SUS AE Source Of Referral For Outpatients")]
-[SourceQuery("SusAESourceOfReferralForOutpatients.xml")]
-internal class SusAESourceOfReferralForOutpatientsRecord
+[Description("SUS AE Source Of Referral For AE")]
+[SourceQuery("SusAESourceOfReferralForAE.xml")]
+internal class SusAESourceOfReferralForAERecord
 {
     public string? NHSNumber { get; set; }
     public string? GeneratedRecordIdentifier { get; set; }
