+++ OmopTransformer/SUS/APC/Observation/ReferralReceivedDateForInpatients/SusAPCReferralReceivedDateForInpatients.cs
@@ -2,14 +2,14 @@
 using OmopTransformer.Omop.Observation;
 using OmopTransformer.Transformation;
 
-namespace OmopTransformer.SUS.APC.Observation.ReferralReceivedDateForOutpatients;
+namespace OmopTransformer.SUS.APC.Observation.ReferralReceivedDateForInpatients;
 
 [Notes(
     "Notes",
     "Observations do not require a standardized test or other activity to generate clinical fact. Typical observations are medical history, family history, lifestyle choices, healthcare utilization patterns, social circumstances etc",
     "Valid Observation Concepts are not enforced to be from any domain.  They should still be standard concepts and typically belong to the Observation or Measurement domain.",
     "Observations can be stored as attribute value pairs, with the attribute as the Observation Concept and the value representing the clinical fact. This fact can be stored as a Concept (value_as_concept), a numerical value (value_as_number) or a verbatim string (value_as_string)")]
-internal class SusAPCReferralReceivedDateForOutpatients : OmopObservation<SusAPCReferralReceivedDateForOutpatientsRecord>
+internal class SusAPCReferralReceivedDateForInpatients : OmopObservation<SusAPCReferralReceivedDateForInpatientsRecord>
 {
     [CopyValue(nameof(Source.NHSNumber))]
     public override string? nhs_number { get; set; }
