+++ OmopTransformer/SUS/APC/Observation/ReferralReceivedDateForInpatients/SusAPCReferralReceivedDateForInpatients.xml
