+++ OmopTransformer/SUS/APC/Observation/ReferralReceivedDateForInpatients/SusAPCReferralReceivedDateForInpatientsRecord.cs
@@ -1,11 +1,11 @@
 ﻿using OmopTransformer.Annotations;
 
-namespace OmopTransformer.SUS.APC.Observation.ReferralReceivedDateForOutpatients;
+namespace OmopTransformer.SUS.APC.Observation.ReferralReceivedDateForInpatients;
 
 [DataOrigin("SUS")]
 [Description("SUS APC Referral Received Date For Outpatients")]
-[SourceQuery("SusAPCReferralReceivedDateForOutpatients.xml")]
-internal class SusAPCReferralReceivedDateForOutpatientsRecord
+[SourceQuery("SusAPCReferralReceivedDateForInpatients.xml")]
+internal class SusAPCReferralReceivedDateForInpatientsRecord
 {
     public string? NHSNumber { get; set; }
     public string? GeneratedRecordIdentifier { get; set; }
